% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_params}
\alias{get_params}
\title{Scatter/line parameters}
\usage{
get_params(x, y, legend = 0)
}
\arguments{
\item{x}{Vector of x values for graph}

\item{y}{Vector of y values for graph}

\item{legend}{Does the graph need a legend? Defaults to 0, replace with 1.}
}
\description{
This sets all parameters for scatter/line plots
}
\examples{
get_params()
}

