% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_params_bar}
\alias{get_params_bar}
\title{Scatter/line parameters}
\usage{
get_params_bar(x, y, legend)
}
\arguments{
\item{x}{Vector of x values for graph}

\item{y}{Vector of y values for graph}

\item{legend}{Does the graph need a legend? Defaults to 0, replace with 1.}
}
\description{
This sets all parameters for bar charts
}
\examples{
get_params_bar()
}

