% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_params_hist}
\alias{get_params_hist}
\title{Histogram parameters}
\usage{
get_params_hist(x, limit_x)
}
\arguments{
\item{x}{Vector of x values for graph}

\item{limit_x}{highest value for x, defaults to 0, which lets ggplot pick}
}
\description{
This sets all parameters for histograms
}
\examples{
get_params_hist()
}

