% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{theme_nyloncalc}
\alias{theme_nyloncalc}
\title{Theme Function}
\usage{
theme_nyloncalc(base_size = 12, base_family = "")
}
\description{
This sets all the ggplot parameters for theme_nyloncalc
}
\examples{
theme_nyloncalc()
}
\keyword{ggplot2}
\keyword{theme}

